requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.6
pydantic-settings==2.7.1
openai==1.63.0
Pillow==11.1.0
pillow-heif==0.21.0
pytest==8.3.4
google-cloud-storage==2.19.0
eval_type_backport==0.2.2
